Former-commit-id: 139c5507c44dddcbbf3fe12eda15c53571d005b6 [formerly f19a241bcc11c3d8a19fb0fbf957f1c8515a9054]
Former-commit-id: f4700f9085b3ae97eb19cc611d37c865cffec21c

+++ Quaver/src/GameState/Gameplay/ScoreManager.cs
@@ -172,7 +172,7 @@ namespace Quaver.GameState.Gameplay
             }
 
             //Update Score todo: actual score calculation
-            ScoreTotal = (int)(1000000 * ((float)ScoreCount / ScoreMax));
+            ScoreTotal = (int)(1000000 * ((double)ScoreCount / ScoreMax));
             //Logger.Log("Score Count: " + ScoreCount + "     Max: " + ScoreMax + "    Note: "+JudgeCount+"/"+ncount, LogColors.GameInfo);
         }
